+++ tagger/src/utils.rs
@@ -447,9 +447,19 @@ impl From<u64> for BitFlags {
 
 impl From<BitFlags> for Vec<i8> {
     fn from(value: BitFlags) -> Self {
-        let mut res = Vec::new();
+        let mut res = Vec::with_capacity(64);
         for i in 0..64 {
             res.push((value.inner & (1 << i) != 0) as i8);
+        }
+        res
+    }
+}
+
+impl From<BitFlags> for Vec<f64> {
+    fn from(value: BitFlags) -> Self {
+        let mut res = Vec::with_capacity(64);
+        for i in 0..64 {
+            res.push((value.inner & (1 << i) != 0) as i8 as f64);
         }
         res
     }